Update README.md
Update to misfitParam

Added a getMisfitParam function that is able to construct misfit param
on the remote worker while using the mesh 2 mesh and pfor already
defined on that worker. This hopefully removes the need for a similar
function for global2local

further update to misparam to support single mesh inversion

Changes in misfitParams

Added type to sigmaBack to prevent ambiguity in the definitions in misfitParam.

updates to misfitParam + doc

Update misfitParam.jl

Added testMisfitParams. Changed dummyMod to identityMod.

Code update for default Mesh2Mesh.

bug fix in misfitParam

bug fix that wasn't updated in the forked repo

Updated misfitParam test to support workers, and identityMod to have UniformScaling